backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
openai==1.82.0
anthropic==0.52.0
google-genai==1.16.1